Moved some short functions back into fn_ctxt.rs
